+++ Sources/MapboxNavigation/NavigationMapView.swift
@@ -1084,6 +1084,15 @@ open class NavigationMapView: UIView {
               from previousLocation: CLLocation? = nil,
               to location: CLLocation,
               animated: Bool = false) {
+        // If the point is outside of the bounds of `MapView` - hide user course view.
+        let point = mapView.mapboxMap.point(for: location.coordinate)
+        if point.x == -1.0 && point.y == -1.0 {
+            userCourseView.isHidden = true
+            return
+        } else {
+            userCourseView.isHidden = false
+        }
+        
         if let previousLocation = previousLocation {
             let point = mapView.mapboxMap.point(for: previousLocation.coordinate)
             userCourseView.center = point